Add files via upload
Forward and Backward approaches are good enough to forecast y values that explain up to 90% of the variation in data, and the model is significant. We received the same single model with three variables from the forward. As a result, we obtained one model in the forward technique and two models in a backward way.